Домашнее задание №13. Приложение с IoC контейнером. Конфигурация DB вынесена в отдельный класс DbConfig.java Исправлен модификатор на final у поля AdminPanelController.userDBExecutor
